requirements.txt
Flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
google-generativeai
markdown
re
jinja2==3.1.2